# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
        "/": {},
        "/api/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Analyze product by name",
                "operationId": "01daa28d8059ba27afbbfad4078d872b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "Neutrogena Hydro Boost"
                                    },
                                    "skin_type": {
                                        "type": "string",
                                        "example": "oily"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product analyzed successfully"
                    },
                    "422": {
                        "description": "Skin profile incomplete"
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/image": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Analyze product by image",
                "operationId": "7ae9026238c67e30f3e9cdb35ea7e72a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "skin_type": {
                                        "type": "string",
                                        "example": "dry"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image analyzed successfully"
                    },
                    "422": {
                        "description": "Skin profile incomplete"
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/history": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get product check history",
                "operationId": "5a93a682e696053986c358c15c2d3c42",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "compatibility",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product checks"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/history/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get single product check",
                "operationId": "28a923ec18113d1b4099dd9842d894c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product check details"
                    },
                    "404": {
                        "description": "Record not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product check",
                "operationId": "c5d46ca8b92fbec3423b67df9bf61303",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Record not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Product",
            "description": "Product"
        }
    ]
}